PD-003 - adding cocktails view

+++ src/PourDecisions.Core/Entities/CocktailIngredient.cs
@@ -5,7 +5,7 @@ namespace PourDecisions.Core.Entities;
 public class CocktailIngredient
 {
     public int Id { get; set; }
-    public decimal AmountValue { get; set; }
+    public int AmountValue { get; set; }
     public AmountUnit AmountUnit { get; set; }
     public bool IsOptional { get; set; }
